add follow-storage
