+++ client/src/services/index.ts
@@ -1,9 +1,16 @@
 export * from './api-adapter';
+
 export * from './chat-adapter';
-export * from './follow-adapter';
+
 export * from './item-adapter';
-export * from './auth-adapter';
+
 export * from './storage-adapter';
+
+export * from './auth-adapter';
+export * from './auth-token-storage';
+
 export * from './user-adapter';
 export * from './user-storage';
-export * from './auth-token-storage';
+
+export * from './follow-adapter';
+export * from './follow-storage';
